Move table toolbar components into table component
